# rework OffsetArrays constructor