Unfortunately, the current version introduces a deeply hidden ambiguity and it blocks https://github.com/JuliaLang/julia/pull/37629

```julia
julia> using Test, OffsetArrays

julia> detect_ambiguities(OffsetArrays)
Tuple{Method,Method}[]

julia> detect_ambiguities(Core, Base, OffsetArrays)
1-element Array{Tuple{Method,Method},1}:
 ((::Type{OffsetArray})(A::AbstractArray, inds::Tuple{Vararg{Union{Colon, AbstractUnitRange, CartesianIndices},N} where N}) in OffsetArrays at /Users/jc/.julia/packages/OffsetArrays/ikTNJ/src/OffsetArrays.jl:135, (::Type{OffsetArray})(A::OffsetArray, offsets::Tuple{Vararg{Int64,N}}) where N in OffsetArrays at /Users/jc/.julia/packages/OffsetArrays/ikTNJ/src/OffsetArrays.jl:155)
```

This PR simplifies the dispatching routes in constructors to fix this ambiguity, and also add missing supports to `OffsetVector`/`OffsetMatrix`.

This PR also renames some of the internal low level utilities with a `_` prefix and moves them to a new file `utils.jl`.

Unsure how this works out, more tests and benchmarks are needed. invalidation test is also needed.