@jishnub 2020-09-18
I'm surprised that the ambiguity doesn't show up in `detect_ambiguities(OffsetArrays)`