@jishnub 2020-09-18
This seems to catch the ambiguity in master, maybe we should switch to this?

```julia
julia> cons = collect(methods(OffsetArray));

julia> any(Base.isambiguous(c1, c2) for c1 in cons, c2 in cons)
true
```