@johnnychen94 2020-09-18
> maybe we should switch to this?

It's okay, we could wait for Aqua v0.5 to catch this: https://github.com/JuliaTesting/Aqua.jl/issues/24