@johnnychen94 2020-09-20
Unfortunately, this still not ready because of the performance issue.

```julia
julia> @btime OffsetArray($a, -2:2, -2:2, -2:2, -2:2); # this PR
  225.305 ns (2 allocations: 128 bytes)

julia> @btime OffsetArray($a, -2:2, -2:2, -2:2, -2:2); # master
   16.556 ns (0 allocations: 0 bytes)
```

I'll need a few more days to investigate it. (Probably this Sunday)