@jishnub 2020-09-21
I realized that `_uncolonindices` does something similar to `Base.to_indices`. I wonder if we might be able to rewrite using this?

```julia
julia> inds1 = OffsetArrays._uncolonindices(zeros(3,4), (2:4, :))
(2:4, Base.OneTo(4))

julia> inds2 = to_indices(zeros(3,4), (2:4, :))
(2:4, Base.Slice(Base.OneTo(4)))

julia> OffsetArray(zeros(3,4), inds1)
3×4 OffsetArray(::Array{Float64,2}, 2:4, 1:4) with eltype Float64 with indices 2:4×1:4:
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

julia> OffsetArray(zeros(3,4), inds2)
3×4 OffsetArray(::Array{Float64,2}, 2:4, 1:4) with eltype Float64 with indices 2:4×1:4:
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
```