@jishnub 2020-09-21
This seems a bit more complicated than I had anticipated. Will look into this after this PR is merged.