@johnnychen94 2020-09-21
The benchmark looks good now:

```julia
using BenchmarkTools, OffsetArrays

a = zeros(5, 5);
@btime OffsetArray{Float64, 2}(undef, -2:2, -2:2);
@btime OffsetArray{Float64, 2}(undef, CartesianIndex(-2, -2):CartesianIndex(2, 2));
@btime OffsetArray($a, -3, -3);
@btime OffsetArray($a, -2:2, :);
@btime OffsetArray($a, CartesianIndex(-2, -2):CartesianIndex(2, 2));

# This PR
# 37.469 ns (1 allocation: 288 bytes)
# 36.476 ns (1 allocation: 288 bytes)
# 1.696 ns (0 allocations: 0 bytes)
# 4.308 ns (0 allocations: 0 bytes)
# 13.447 ns (0 allocations: 0 bytes)

# master
# 31.802 ns (1 allocation: 288 bytes)
# 40.283 ns (1 allocation: 288 bytes)
# 1.421 ns (0 allocations: 0 bytes)
# 4.313 ns (0 allocations: 0 bytes)
# 13.080 ns (0 allocations: 0 bytes)
```