@johnnychen94 2020-09-21
Most of the previous test cases are rewritten more compactly. I've tested it locally that the previous test also passes.

I think this one is ready.  In case there are further comments, I'll leave it one day or two before I merge this so that we could continue #157 and #147